Transformo el id byte de Categpria a integer para solucionar problemas con BBDD

+++ Catalogo/src/main/java/com/catalogo/DAO/CategoryDAO.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import com.catalogo.domain.Category;
 
 @Repository("CategoryDAO")
-public interface CategoryDAO extends CrudRepository<Category,Byte>{
+public interface CategoryDAO extends CrudRepository<Category,Integer>{
 
 }
